+++ src/orchard/java/source_files.clj
@@ -74,7 +74,12 @@
 (defn- combine-archive-url ^URL [^File archive, relative-filename]
   ;; Even though the JDK stores sources in a zip archive, we still use this
   ;; function that prefixes the URL with jar:. This is fine.
-  (io/as-url (format "jar:file:%s!/%s" archive relative-filename)))
+  (io/as-url (format "jar:file:%s!/%s"
+                     ;; Replace backslashes with forward slashes. Needed on
+                     ;; Windows. Forward slashes would still work there, while
+                     ;; backslashes are forbidden in URLs.
+                     (string/replace archive "\\" "/")
+                     relative-filename)))
 
 (defn- locate-source-url-in-jdk-sources
   "Try to find the source file for `klass` in sources included with JDK."